requirements.txt
gitdb==4.0.9
GitPython==3.1.37
smmap==5.0.0
typing-extensions==3.10.0.2
